[not verified] Protect: Move vuls accordion list to own file
